refactor: merge upgrade (xatp & align gas price)
