--- app/upgrades/align_gas_price/const.go
@@ -1,5 +1,0 @@
-package align_gas_price
-
-const (
-	UpgradeName = "AlignGasPrice"
-)
