+++ app/upgrades/xatp/constants.go
@@ -1,5 +1,13 @@
-package v2
+package xatp
+
+import (
+	xatptypes "github.com/xpladev/xpla/x/xatp/types"
+)
 
 const (
 	UpgradeName = "XATP"
+)
+
+var (
+	AddModules = []string{xatptypes.ModuleName}
 )
