+++ app/upgrades/xatp/upgrades.go
@@ -1,4 +1,4 @@
-package v2
+package xatp
 
 import (
 	"encoding/json"
@@ -11,26 +11,33 @@ import (
 
 	xatpKeeper "github.com/xpladev/xpla/x/xatp/keeper"
 	xatptypes "github.com/xpladev/xpla/x/xatp/types"
+
+	feemarketkeeper "github.com/evmos/ethermint/x/feemarket/keeper"
 )
 
-// CreateUpgradeHandler creates an SDK upgrade handler for v2
+// CreateUpgradeHandler creates an SDK upgrade handler for xatp
 func CreateUpgradeHandler(
 	mm *module.Manager,
 	configurator module.Configurator,
 	xatpKeeper xatpKeeper.Keeper,
+	fk feemarketkeeper.Keeper,
 ) upgradetypes.UpgradeHandler {
 	return func(ctx sdk.Context, plan upgradetypes.Plan, vm module.VersionMap) (module.VersionMap, error) {
 
 		vm[xatptypes.ModuleName] = 1
 
-		var params xatptypes.Params
+		var msg UpgradeXatpMsg
 
-		err := json.Unmarshal([]byte(plan.Info), &params)
+		err := json.Unmarshal([]byte(plan.Info), &msg)
 		if err != nil {
 			return nil, sdkerrors.Wrapf(sdkerrors.ErrJSONUnmarshal, UpgradeName+": unmarshal error")
 		}
 
-		xatpKeeper.SetParams(ctx, params)
+		xatpKeeper.SetParams(ctx, msg.XATP)
+
+		feemarketParams := fk.GetParams(ctx)
+		feemarketParams.MinGasPrice = msg.MinGasPrice
+		fk.SetParams(ctx, feemarketParams)
 
 		return mm.RunMigrations(ctx, configurator, vm)
 	}
